some refactoring to separate the concerns in Mongo::Retryable
